ResumeOrbit/backend/requirements.txt
spacy>=3.8.0
Flask>=3.0.0
Flask-CORS>=4.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
Werkzeug>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.15.2
